implement time intervals in the time tracker
The format of tracker's csv files has been extended to include an
intervals field.
This allows the tracker to be resumed automatically if it was
stopped abruptly.
Also, the stats view is updated to be able to show time intervals.

I also snuck in some other smaller things (too lazy to make separate
commits):
- Dual search of tasks and projects in the stats view.
- Some bug fixes.
- Lots of code refactored.
- Keyboard shortcuts can be triggerd when the popup menu is open.
- Add a shortcut for opening the current todo.txt file, etc...
- Stuff..