Remove actor property which has been removed.
Note that there is other references to .actor in the files. I'm not sure if they are ok.
Since it's javascript, maybe the bugs will only trigger if the code is actually executed.

I cannot find any decent documentation on gnome-shell, nor a way to debug it without
log-out after every change to the sources, so I haven't tried going any deeper.